Send latitude and longitude data as fixed point number

+++ MIDAS/src/sensor_data.h
@@ -104,8 +104,8 @@ struct Voltage {
 };
 
 struct GPS {
-    float latitude = 0;
-    float longitude = 0;
+    int32_t latitude = 0;
+    int32_t longitude = 0;
     float altitude = 0;
     float speed = 0;
     uint16_t satellite_count = 0;
